Create package.json
The first goal is to setup a simple HTML webpage
that serves out a form and a list of messages.
We’re going to use the Node.JS web framework `express`
to this end.

So as a first step, let’s create a package.json
manifest file that describes our project.

I created this file using `npm init` and used the
default values except for "name" and "author".