Emit event from index.html
The main idea behind Socket.IO is that you
can send and receive any events you want,
with any data you want. Any objects that can
be encoded as JSON will do, and binary data
is supported too.

Let’s edit the script section in index.html
so that when the user types in a message, the
server gets it as a chat message event.